Introduce Event wrapper
This mostly abstracts away the use of libical and removes the EventPrivate
structure which abused the libical API. Extra data members, such as URL and
ETag which are used for CalDAV can now be stored simply in the new structure.

This commit also removes the distinction between saving and updating events.
If a Calendar class makes a distinction, it will be responsible for checking
whether the added Event already exists.

Finally, the way events are saved in LocalCalendar is changed to clone them
into the events GSList rather than retaining a pointer to the original. This
is still experimental, but works toward individual saving/loading of events.