resolves #72: incremental sync for outlook calendars
This changes the mechanism for outlook sync to be based around
calendarView. Previously, we tried to fetch all the events in
the calendar (which does not provide full occurrence info), then
fetch the recurrence info for the current view span. Unfortunately,
this method cannot be extended to incremental sync. Instead, take
the approach of synchronizing only the current view span, which
does provide an incremental sync method. A full sync per span
happens when the span is changed, but the actual events and
recurrences are preserved. Since the aim is to minimize the size
of the request for automatic sync, this is more important than
supporting incremental sync across different view spans.

Also:
- move to the new graph API endpoint
- add a conversion map from Windows timezone names