resolves #51: wrong week shown on startup
This issue was originally reported as showing the the next week when the current
day is Sunday, but it turned out to be more complex when ical_week_number was
found not to return the week number according to ISO 8601. This commit implements
routines to correctly figure out the current week and fixes the "wrong week" bugs.